requirements.txt
alembic==1.17.1
anyio==4.12.1
APScheduler==3.10.4
black==26.1.0
certifi==2026.1.4
click==8.3.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-telegram-bot==22.6
pytokens==0.4.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
